refactor: formatting form
