+++ my-app/src/components/Form/Form.props.ts
@@ -1,4 +1,4 @@
-import { IFormCard } from 'components/FormCard/FormCard.interface';
+import { IFormCard } from 'components/Form/FormCard/FormCard.interface';
 import { DetailedHTMLProps, HTMLAttributes } from 'react';
 
 export interface FormProps
