+++ my-app/src/components/Form/FormList/FormList.module.css
@@ -1,0 +1,9 @@
+.cardList {
+	display: grid;
+	grid-template-columns: repeat( auto-fit, minmax(300px, 1fr) );
+	align-items: center;
+	justify-items: center;
+	padding: 0 20px;
+	gap: 30px 20px;
+
+}
