+++ my-app/src/components/Form/FormList/FormList.props.ts
@@ -1,0 +1,7 @@
+import { IFormCard } from 'components/Form/FormCard/FormCard.interface';
+import { HTMLAttributes, DetailedHTMLProps } from 'react';
+
+export interface FormListProps
+  extends DetailedHTMLProps<HTMLAttributes<HTMLDivElement>, HTMLDivElement> {
+  data: IFormCard[];
+}
