--- my-app/src/dataBase/catalog.interface.ts
@@ -1,17 +1,0 @@
-export interface ICatalog {
-  id: string;
-  brand: string;
-  name: string;
-  image: string;
-  price: string;
-  year: string;
-  size: string;
-  resolution: string;
-  aspectRatio: string;
-  refRate: string;
-  count: string;
-  portrait: string;
-  gaming: string;
-  pro: string;
-  favorite: string;
-}
