--- my-app/src/dataBase/catalog.ts
@@ -1,345 +1,0 @@
-import { ICatalog } from './catalog.interface';
-
-export const CATALOG: ICatalog[] = [
-  {
-    id: 'el1',
-    brand: 'apple',
-    name: 'Pro Display XDR',
-    image: './img/content/apple/apple.jpeg',
-    price: '14530.00',
-    year: '2019',
-    size: '32',
-    resolution: '6016x3384',
-    aspectRatio: '16:9',
-    refRate: '60',
-    count: '1',
-    portrait: 'false',
-    gaming: 'false',
-    pro: 'true',
-    favorite: 'false',
-  },
-  {
-    id: 'el2',
-    brand: 'dell',
-    name: 'U3219Q',
-    image: './img/content/dell/Dell_U3219Q.jpeg',
-    price: '3095.09',
-    year: '2018',
-    size: '31.5',
-    resolution: '3840x2160',
-    aspectRatio: '16:9',
-    refRate: '60',
-    count: '3',
-    portrait: 'false',
-    gaming: 'false',
-    pro: 'true',
-    favorite: 'true',
-  },
-  {
-    id: 'el3',
-    brand: 'dell',
-    name: 'UltraSharp U4021QW',
-    image: './img/content/dell/dell_ultraSharp_u4021qw.jpeg',
-    price: '6499.00',
-    year: '2021',
-    size: '40',
-    resolution: '5120x2160',
-    aspectRatio: '21:9',
-    refRate: '60',
-    count: '2',
-    portrait: 'false',
-    gaming: 'false',
-    pro: 'true',
-    favorite: 'false',
-  },
-  {
-    id: 'el4',
-    brand: 'dell',
-    name: 'UltraSharp U4919DW',
-    image: './img/content/dell/Dell_UltraSharp_U4919DW.jpeg',
-    price: '5399.00',
-    year: '2018',
-    size: '49',
-    resolution: '5120x1440',
-    aspectRatio: '32:9',
-    refRate: '60',
-    count: '3',
-    portrait: 'false',
-    gaming: 'false',
-    pro: 'true',
-    favorite: 'false',
-  },
-  {
-    id: 'el5',
-    brand: 'dell',
-    name: 'UltraSharp UP3221Q',
-    image: './img/content/dell/dell_ultraSharp_up3221q.jpeg',
-    price: '17056.78',
-    year: '2020',
-    size: '31.5',
-    resolution: '3840x2160',
-    aspectRatio: '16:9',
-    refRate: '60',
-    count: '1',
-    portrait: 'true',
-    gaming: 'false',
-    pro: 'true',
-    favorite: 'true',
-  },
-  {
-    id: 'el6',
-    brand: 'HP',
-    name: 'M34D 3B1W4AA',
-    image: './img/content/hp/hp_m34d_3b1w4aa.jpeg',
-    price: '1950.00',
-    year: '2021',
-    size: '34',
-    resolution: '3440x1440',
-    aspectRatio: '21:9',
-    refRate: '100',
-    count: '5',
-    portrait: 'false',
-    gaming: 'false',
-    pro: 'true',
-    favorite: 'false',
-  },
-  {
-    id: 'el7',
-    brand: 'HP',
-    name: 'U28',
-    image: './img/content/hp/hp_u28.jpeg',
-    price: '2478.78',
-    year: '2021',
-    size: '28',
-    resolution: '3840x2160',
-    aspectRatio: '16:9',
-    refRate: '60',
-    count: '2',
-    portrait: 'true',
-    gaming: 'false',
-    pro: 'true',
-    favorite: 'false',
-  },
-  {
-    id: 'el8',
-    brand: 'LG',
-    name: '32Ul750-W',
-    image: './img/content/lg/lg_32ul750-w.jpeg',
-    price: '2149.43',
-    year: '2022',
-    size: '31.5',
-    resolution: '3840x2160',
-    aspectRatio: '16:9',
-    refRate: '60',
-    count: '4',
-    portrait: 'false',
-    gaming: 'false',
-    pro: 'true',
-    favorite: 'true',
-  },
-  {
-    id: 'el9',
-    brand: 'LG',
-    name: 'UltraFine 32EP950-B',
-    image: './img/content/lg/lg_ultraFine_32ep950-b.jpeg',
-    price: '8688.00',
-    year: '2020',
-    size: '31.5',
-    resolution: '3840x2160',
-    aspectRatio: '16:9',
-    refRate: '60',
-    count: '2',
-    portrait: 'true',
-    gaming: 'false',
-    pro: 'true',
-    favorite: 'false',
-  },
-  {
-    id: 'el10',
-    brand: 'LG',
-    name: 'UltraWide 40WP95C-W',
-    image: './img/content/lg/lg_ultraWide_40wp95c-w.jpeg',
-    price: '5799.00',
-    year: '2022',
-    size: '39.7',
-    resolution: '5120x2160',
-    aspectRatio: '21:9',
-    refRate: '75',
-    count: '2',
-    portrait: 'false',
-    gaming: 'false',
-    pro: 'true',
-    favorite: 'false',
-  },
-  {
-    id: 'el11',
-    brand: 'LG',
-    name: 'UltraWide 49W195C-W',
-    image: './img/content/lg/lg_ultraWide_49wl95c-w.jpeg',
-    price: '4880.00',
-    year: '2021',
-    size: '49',
-    resolution: '5120x1440',
-    aspectRatio: '32:9',
-    refRate: '60',
-    count: '1',
-    portrait: 'false',
-    gaming: 'false',
-    pro: 'true',
-    favorite: 'false',
-  },
-  {
-    id: 'el12',
-    brand: 'Samsung',
-    name: 'S34J550WQI',
-    image: './img/content/samsung/samsung_s34j550wqi.jpeg',
-    price: '1571.28',
-    year: '2019',
-    size: '34.1',
-    resolution: '3440x1440',
-    aspectRatio: '21:9',
-    refRate: '75',
-    count: '3',
-    portrait: 'false',
-    gaming: 'true',
-    pro: 'false',
-    favorite: 'false',
-  },
-  /**=============================================gaming============================ */
-  {
-    id: 'el13',
-    brand: 'LG',
-    name: 'UltraGear 38GN950-B',
-    image: './img/content/lg/LG_UltraGear_38GN950-B.jpeg',
-    price: '6161.38',
-    year: '2020',
-    size: '37.5',
-    resolution: '3840x1600',
-    aspectRatio: '21:9',
-    refRate: '160',
-    count: '1',
-    portrait: 'false',
-    gaming: 'true',
-    pro: 'false',
-    favorite: 'false',
-  },
-  {
-    id: 'el14',
-    brand: 'LG',
-    name: '27GN750-B',
-    image: './img/content/lg/LG_27GN750-B.jpeg',
-    price: '4685.05',
-    year: '2020',
-    size: '27',
-    resolution: '1920x1080',
-    aspectRatio: '16:9',
-    refRate: '240',
-    count: '1',
-    portrait: 'false',
-    gaming: 'true',
-    pro: 'false',
-    favorite: 'true',
-  },
-  {
-    id: 'el15',
-    brand: 'Dell',
-    name: 'AW2521HF',
-    image: './img/content/dell/Dell_AW2521HF.jpeg',
-    price: '4175.71',
-    year: '2020',
-    size: '24.5',
-    resolution: '1920x1080',
-    aspectRatio: '16:9',
-    refRate: '240',
-    count: '1',
-    portrait: 'false',
-    gaming: 'true',
-    pro: 'false',
-    favorite: 'false',
-  },
-  {
-    id: 'el16',
-    brand: 'Samsung',
-    name: 'Odyssey G3 LC32G35TFQIXCI',
-    image: './img/content/samsung/Samsung_Odyssey_G3_LC32G35TFQIXCI.jpeg',
-    price: '2338.60',
-    year: '2022',
-    size: '32',
-    resolution: '1920x1080',
-    aspectRatio: '16:9',
-    refRate: '165',
-    count: '2',
-    portrait: 'false',
-    gaming: 'true',
-    pro: 'false',
-    favorite: 'false',
-  },
-  {
-    id: 'el17',
-    brand: 'HP',
-    name: 'OMEN 27i 8AC94AA',
-    image: './img/content/hp/HP_OMEN_27i_8AC94AA.jpeg',
-    price: '2010.00',
-    year: '2020',
-    size: '27',
-    resolution: '2560x1440',
-    aspectRatio: '16:9',
-    refRate: '165',
-    count: '5',
-    portrait: 'false',
-    gaming: 'true',
-    pro: 'false',
-    favorite: 'false',
-  },
-  {
-    id: 'el18',
-    brand: 'Dell',
-    name: 'G2422Hs',
-    image: './img/content/dell/Dell_G2422Hs.jpeg',
-    price: '1399.00',
-    year: '2022',
-    size: '37.5',
-    resolution: '1920x1080',
-    aspectRatio: '16:9',
-    refRate: '165',
-    count: '3',
-    portrait: 'false',
-    gaming: 'true',
-    pro: 'false',
-    favorite: 'false',
-  },
-  {
-    id: 'el19',
-    brand: 'Samsung',
-    name: 'S34J550WQI',
-    image: './img/content/samsung/samsung_s34j550wqi.jpeg',
-    price: '1210.90',
-    year: '2021',
-    size: '27',
-    resolution: '1920x1080',
-    aspectRatio: '16:9',
-    refRate: '144',
-    count: '3',
-    portrait: 'false',
-    gaming: 'true',
-    pro: 'false',
-    favorite: 'false',
-  },
-  {
-    id: 'el20',
-    brand: 'HP',
-    name: 'X27qc',
-    image: './img/content/hp/HP_X27qc.jpeg',
-    price: '1299.00',
-    year: '2021',
-    size: '27',
-    resolution: '2560x1440',
-    aspectRatio: '16:9',
-    refRate: '165',
-    count: '4',
-    portrait: 'false',
-    gaming: 'true',
-    pro: 'false',
-    favorite: 'true',
-  },
-];
