+++ my-app/src/pages/FormPage/FormPage.state.ts
@@ -1,4 +1,4 @@
-import { IFormCard } from 'components/FormCard/FormCard.interface';
+import { IFormCard } from 'components/Form/FormCard/FormCard.interface';
 
 export interface FormPageState {
   cardList: IFormCard[] | [];
