+++ my-app/src/pages/FormPage/FormPage.test.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
-import { fireEvent, render, screen } from '@testing-library/react';
+import { render, screen } from '@testing-library/react';
 import { MemoryRouter } from 'react-router-dom';
 import FormPage from './FormPage';
-import userEvent from '@testing-library/user-event';
+// import userEvent from '@testing-library/user-event';
 
 describe('FormPage', () => {
   test('FormPage render', () => {
@@ -14,4 +14,35 @@ describe('FormPage', () => {
     expect(screen.getByTestId('form')).toBeInTheDocument();
     expect(screen.queryByTestId('FormCard')).not.toBeInTheDocument();
   });
+  // test('add card', () => {
+  //   render(
+  //     <MemoryRouter initialEntries={['/form']}>
+  //       <FormPage />
+  //     </MemoryRouter>
+  //   );
+  //   const name = screen.getByTestId('name');
+  //   const surname = screen.getByTestId('surname');
+  //   const zipCode = screen.getByTestId('zipCode');
+  //   const country = screen.getByTestId('country');
+  //   const birthday = screen.getByTestId('birthday');
+  //   const agree = screen.getByTestId('agree');
+  //   const button = screen.getByRole('button');
+  //   fireEvent.input(name, { target: { value: 'test' } });
+  //   fireEvent.input(surname, { target: { value: 'tester' } });
+  //   fireEvent.input(zipCode, { target: { value: '1234567' } });
+  //   fireEvent.change(country, { target: { value: 'Belarus' } });
+  //   fireEvent.input(birthday, { target: { value: '2000-02-02' } });
+  //   userEvent.click(agree);
+  // fireEvent.click(button);
+  // fireEvent.input(name, { target: { value: 'test' } });
+  // fireEvent.input(surname, { target: { value: 'test' } });
+  // fireEvent.input(zipCode, { target: { value: '1234567' } });
+  // fireEvent.change(country, { target: { value: 'Belarus' } });
+  // fireEvent.input(birthday, { target: { value: '2000-02-02' } });
+  // userEvent.click(agree);
+  // fireEvent.click(button);
+  // expect(screen.getByText('tester')).toBe('tester');
+  // expect(screen.getByTestId('form')).toBeInTheDocument();
+  // expect(screen.getByTestId('FormCard')).toBeInTheDocument();
+  // });
 });
